refactor(scripts): Convert to TypeScript
