+++ packages/scripts/package.json
@@ -17,8 +17,9 @@
     "yargs": "^17.1.1"
   },
   "scripts": {
-    "collect-i18n": "node ./collect-i18n-messages.js ../**/src/{,**/}*.{j,t}s{,x} ../**/i18n/en-US.yml > i18n.csv",
-    "validate-i18n": "node ./validate-i18n.js ../**/src/{,**/}*.{j,t}s{,x} ../**/i18n/*.yml"
+    "collect-i18n": "node src/collect-i18n-messages.js ../**/src/{,**/}*.{j,t}s{,x} ../**/i18n/en-US.yml > i18n.csv",
+    "tsc": "tsc",
+    "validate-i18n": "node src/validate-i18n.js ../**/src/{,**/}*.{j,t}s{,x} ../**/i18n/*.yml"
   },
   "bugs": {
     "url": "https://github.com/opentripplanner/otp-ui/issues"
